LayoutTests/platform/blackberry/accessibility/aria-option-role-expected.txt

option 1
option 2
This tests that the aria 'option' role works as expected. That is, it becomes a static text element with no children.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL firstChild.role should be AXRole: AXStaticText. Was .
FAIL firstChild.title should be AXTitle: option 1. Was .
FAIL secondChild.role should be AXRole: AXStaticText. Was .
FAIL secondChild.description should be AXDescription: label 2. Was .
PASS firstChild.childrenCount is 0
PASS secondChild.childrenCount is 0
PASS successfullyParsed is true

TEST COMPLETE
